requirements.txt
numpy==2.1.3
sounddevice==0.5.2
soundfile==0.13.1
librosa==0.11.0
noisereduce==3.0.3
tensorflow==2.19.0
keyboard==0.13.5
matplotlib==3.10.3
openai-whisper==20240930
torch==2.7.1
transformer
#python 3.10.0
